+++ dflib-avro/src/test/java/org/dflib/avro/Avro_SchemaSerializationTest.java
@@ -4,7 +4,6 @@ import org.apache.avro.Schema;
 import org.apache.avro.SchemaBuilder;
 import org.dflib.DataFrame;
 import org.dflib.Series;
-import org.dflib.avro.Avro;
 import org.junit.jupiter.api.Test;
 
 import java.io.ByteArrayInputStream;
@@ -66,7 +65,7 @@ public class Avro_SchemaSerializationTest {
                         "{\"name\":\"Double\",\"type\":[\"double\",\"null\"]}," +
                         "{\"name\":\"bool\",\"type\":\"boolean\"}," +
                         "{\"name\":\"Bool\",\"type\":[\"boolean\",\"null\"]}," +
-                        "{\"name\":\"String\",\"type\":[{\"type\":\"string\",\"avro.java.string\":\"String\"},\"null\"]}," +
+                        "{\"name\":\"String\",\"type\":[\"string\",\"null\"]}," +
                         "{\"name\":\"byte_array\",\"type\":[{\"type\":\"bytes\",\"logicalType\":\"dflib-bytearray\"},\"null\"]}," +
                         "{\"name\":\"LocalDate\",\"type\":[{\"type\":\"long\",\"logicalType\":\"dflib-localdate\"},\"null\"]}]}",
                 out.toString(StandardCharsets.UTF_8.name()));