+++ RELEASE-NOTES.md
@@ -11,6 +11,7 @@
 * #332 Compacting CSV columns on load
 * #333 Compacting Avro columns on load
 * #334 AvroLoader: "string" type without "avro.java.string":"String" is processed incorrectly
+* #335 Stop adding "avro.java.string" property to String field schemas
 
 ## Release 1.0.0-M23